requirements.txt
ffmpeg-python 
openai-whisper
googletrans==4.0.0-rc1
transformers
gtts